requirements.txt
asgiref==3.4.1
Django==3.2
django-appconf==1.0.5
django-cors-headers==3.10.1
django-filter==21.1
django-session-timeout==0.1.0
django-stdimage==5.3.0
django-taggit==2.0.0
django-widget-tweaks==1.4.9
Pillow==8.4.0
pyclean==2.0.0
python-decouple==3.5
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
